Large refactorings to database handling.
Instead of the pessimistic locking introduced in the last commit, we can use optimistic locking. This should boost performance in all scenarios. This commit also introduces the base of the logging system, but it doesn't work currently. I still need to figure out how and when to register currencies, and where and when to allow plugins to register custom LogContexts.